% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_index_symbol.R
\name{check_index_symbol}
\alias{check_index_symbol}
\title{Check Confidence Index Symbol}
\usage{
check_index_symbol(symbol)
}
\arguments{
\item{symbol}{A string with a Confidence Index symbol:
\itemize{
\item US1YI: U.S. One Year Index
\item USVI: U.S. Valuation Index
\item USBODI: U.S. Buy-On-Dips Confidence Index
\item USCI: U.S. Crash Confidence Index
}}
}
\value{
A string with the \code{symbol} or an error if it
is not found.
}
\description{
This function validates the symbol input.
}
\examples{
check_index_symbol("USVI")
}
\seealso{
\code{\link{US_confidence_indices}}

\code{\link{US_confidence_indices_url}}

\code{\link{find_table_index}}

\code{\link{get_nodes}}
}
