% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{find_table_index}
\alias{find_table_index}
\title{Extract indices data from html node}
\usage{
find_table_index(nodes, investor = "Institutional")
}
\arguments{
\item{nodes}{A \code{xml_nodeset} class data with indices information}

\item{investor}{A string with the type of investor:
\code{"Institutional"} or \code{"Individual"}}
}
\value{
A \code{data.frame} with ancillary data.
}
\description{
Extract indices data from html node
}
\note{
The default value for \code{investor} is "Institutional".

This function is only for internal use.
}
\examples{
\dontrun{
find_table_index(get_nodes("US1YI"), investor = "Individual")
}

}
\seealso{
\code{\link{get_index}}

\code{\link{get_nodes}}
}
