% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{get_index}
\alias{get_index}
\title{Get confidence index data from Yale School of Management's web}
\usage{
get_index(symbol = "US1YI")
}
\arguments{
\item{symbol}{A string with a Confidence Index symbol:
\itemize{
\item \emph{US1YI}: U.S. One-Year Confidence Index
\item \emph{USVI}: U.S. Valuation Confidence Index
\item \emph{USBODI}: U.S. Buy-On-Dips Confidence Index
\item \emph{USCI}: U.S. Crash Confidence Index
}}
}
\value{
A \code{data.table} with index value, standard deviation, and type of investor, by date.
}
\description{
Stock market confidence indexes are derived from survey data on
the behavior of U.S. investors.
}
\note{
The default value for \code{symbol} is "US1YI".
}
\examples{
\dontrun{
get_index("USCI")
}
}
\references{
\emph{Yale School of Management}, International Center for Finance: \href{https://som.yale.edu/faculty-research/our-centers-initiatives/international-center-finance/data/stock-market-confidence}{Stock Market Confidence Indices}

Stock Market Confidence Indices: \href{https://som.yale.edu/faculty-research/centers-initiatives/international-center-for-finance/data/stock-market-confidence-indices/stock-market-confidence-indices}{Terms of Use}
}
\seealso{
\code{\link{US_confidence_indices}}

\code{\link{US_confidence_indices_url}}

\code{\link{get_index_description}}

\code{\link{get_index_info}}

\code{\link{check_index_symbol}}
}
