% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_info.R
\name{get_index_info}
\alias{get_index_info}
\title{Find Confidence Index information}
\usage{
get_index_info(symbol = "US1YI")
}
\arguments{
\item{symbol}{A string with a Confidence Index symbol:
\itemize{
\item \emph{US1YI}: U.S. One Year Index
\item \emph{USVI}: U.S. Valuation Index
\item \emph{USBODI}: U.S. Buy-On-Dips Confidence Index
\item \emph{USCI}: U.S. Crash Confidence Index
}}
}
\value{
A string with the Confidence Index description
}
\description{
Find Confidence Index information
}
\note{
The default value for \code{symbol} is "US1YI".
}
\examples{
get_index_info("USCI")
}
\seealso{
\code{\link{US_confidence_indices}}

\code{\link{get_index}}

\code{\link{get_index_description}}

\code{\link{check_index_symbol}}
}
