% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{get_nodes}
\alias{get_nodes}
\title{Get html nodes from confidence indices at
Yale School of Management's web}
\usage{
get_nodes(symbol = "US1YI")
}
\arguments{
\item{symbol}{A string with a Confidence Index symbol:
\itemize{
\item US1YI: U.S. One Year Index
\item USVI: U.S. Crash Confidence Index
\item USBODI: U.S. Buy-On-Dips Confidence Index
\item USCI: U.S. Valuation Index
}}
}
\value{
A \code{xml_nodeset} class data with indices information
}
\description{
Get html nodes from confidence indices at
Yale School of Management's web
}
\note{
This function is only for internal use.
}
\examples{
\dontrun{
get_nodes("US1YI")
}
}
\seealso{
\code{\link{get_index}}

\code{\link{find_table_index}}
}
