% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_confidence_indices.R
\name{US_confidence_indices}
\alias{US_confidence_indices}
\title{US confidence indices data.frame}
\usage{
US_confidence_indices()
}
\value{
A \code{data.frame} with symbols and urls from
\strong{Yale School of Management}'s web page.
}
\description{
US confidence indices data.frame
}
\examples{
\dontrun{
US_confidence_indices()
}
}
\references{
\emph{Yale School of Management}, International Center for Finance: \href{https://som.yale.edu/faculty-research/our-centers-initiatives/international-center-finance/data/stock-market-confidence}{Stock Market Confidence Indices}

Stock Market Confidence Indices: \href{https://som.yale.edu/faculty-research/centers-initiatives/international-center-for-finance/data/stock-market-confidence-indices/stock-market-confidence-indices}{Terms of Use}
}
\seealso{
\code{\link{US_confidence_indices_url}}

\code{\link{get_index}}
}
