% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_confidence_indices_url.R
\name{US_confidence_indices_url}
\alias{US_confidence_indices_url}
\title{Find Confidence Index symbol url}
\usage{
US_confidence_indices_url(symbol = "US1YI")
}
\arguments{
\item{symbol}{A string with the Confidence Index symbol}
}
\value{
A string with the \strong{Yale School of Management}'s web
 page symbol url
}
\description{
Find Confidence Index symbol url
}
\examples{
\dontrun{
US_confidence_indices_url("USCI")
}
}
\seealso{
\code{\link{US_confidence_indices}}
}
