% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{get_nodes}
\alias{get_nodes}
\title{Get html nodes from confidence indices at
Yale School of Management's web}
\usage{
get_nodes(symbol = "US1YI")
}
\arguments{
\item{symbol}{A string with a Confidence Index symbol:
\itemize{
\item \emph{US1YI}: U.S. One Year Index
\item \emph{USVI}: U.S. Valuation Index
\item \emph{USBODI}: U.S. Buy-On-Dips Confidence Index
\item \emph{USCI}: U.S. Crash Confidence Index
}}
}
\value{
A \code{xml_nodeset} class data with indices information
}
\description{
Get html nodes from confidence indices at
Yale School of Management's web
}
\note{
The default value for \code{symbol} is "US1YI".

This function is only for internal use.
}
\examples{
\dontrun{
get_nodes("US1YI")
}
}
\seealso{
\code{\link{get_index}}

\code{\link{find_table_index}}

\code{\link{check_index_symbol}}
}
