% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_confidence_indices_url.R
\name{US_confidence_indices_url}
\alias{US_confidence_indices_url}
\title{Get Confidence Index symbol url}
\usage{
US_confidence_indices_url(symbol = "US1YI")
}
\arguments{
\item{symbol}{A string with a Confidence Index symbol:
\itemize{
\item \emph{US1YI}: U.S. One-Year Confidence Index
\item \emph{USVI}: U.S. Valuation Confidence Index
\item \emph{USBODI}: U.S. Buy-On-Dips Confidence Index
\item \emph{USCI}: U.S. Crash Confidence Index
}}
}
\value{
A string with the \strong{Yale School of Management}'s web
page symbol url
}
\description{
Get Confidence Index symbol url
}
\note{
The default value for \code{symbol} is "US1YI".
}
\examples{
\dontrun{
US_confidence_indices_url("USCI")
}
}
\seealso{
\code{\link{US_confidence_indices}}

\code{\link{get_index_description}}

\code{\link{get_index_info}}

\code{\link{get_index}}
}
